refactor: decrypt as function
